+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+[unreleased]
+- **Added** `path` property to error object
+- **Fixed** repeated errors in nested model scenario
+
 [3.3.1] - 2017-12-28
 - **Added** new type `any`
 - **Added** new property `allowEmpty`, now it's possible check if a string or an array or an object are empties