+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-[unreleased]
+[4.0.0] - 2018-01-03
 - **Added** `path` property to error object
 - **Added** `type` property to error object
 - **Added** `detectUnknown` property to model definition